Extract some methods to WebdriverTypeHelper
