+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2016-11-07 18:34:15 +0300 using RuboCop version 0.45.0.
+# on 2016-11-07 18:47:59 +0300 using RuboCop version 0.45.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -13,6 +13,7 @@ Lint/RescueException:
     - 'lib/onlyoffice_webdriver_wrapper/helpers/headless_helper/real_display_tools.rb'
     - 'lib/onlyoffice_webdriver_wrapper/webdriver.rb'
     - 'lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_attributes_helper.rb'
+    - 'lib/onlyoffice_webdriver_wrapper/webdriver/webdriver_type_helper.rb'
 
 # Offense count: 11
 Metrics/AbcSize:
@@ -21,7 +22,7 @@ Metrics/AbcSize:
 # Offense count: 1
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 908
+  Max: 797
 
 # Offense count: 3
 Metrics/CyclomaticComplexity:
@@ -37,6 +38,11 @@ Metrics/LineLength:
 # Configuration parameters: CountComments.
 Metrics/MethodLength:
   Max: 35
+
+# Offense count: 1
+# Configuration parameters: CountComments.
+Metrics/ModuleLength:
+  Max: 112
 
 # Offense count: 3
 # Configuration parameters: CountKeywordArgs.
